+++ NINA/PlateSolving/PlateSolveParameter.cs
@@ -1,49 +1,44 @@
-﻿using NINA.Utility.Astrometry;
+﻿#region "copyright"
+
+/*
+    Copyright © 2016 - 2019 Stefan Berg <isbeorn86+NINA@googlemail.com>
+
+    This file is part of N.I.N.A. - Nighttime Imaging 'N' Astronomy.
+
+    N.I.N.A. is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    N.I.N.A. is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with N.I.N.A..  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#endregion "copyright"
+
+using NINA.Utility.Astrometry;
 using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NINA.PlateSolving {
 
     internal class PlateSolveParameter {
         public double FocalLength { get; set; }
         public double PixelSize { get; set; }
-        public double ImageWidth { get; set; }
-        public double ImageHeight { get; set; }
         public double SearchRadius { get; set; }
         public double Regions { get; set; }
         public int DownSampleFactor { get; set; }
         public int MaxObjects { get; set; }
         public Coordinates Coordinates { get; set; }
-        public Stream Image { get; set; }
-
-        public double ArcSecPerPixel {
-            get {
-                return Astrometry.ArcsecPerPixel(PixelSize, FocalLength);
-            }
-        }
-
-        public double FoVH {
-            get {
-                return Astrometry.ArcminToDegree(Astrometry.FieldOfView(ArcSecPerPixel, ImageHeight));
-            }
-        }
-
-        public double FoVW {
-            get {
-                return Astrometry.ArcminToDegree(Astrometry.FieldOfView(ArcSecPerPixel, ImageWidth));
-            }
-        }
 
         public override string ToString() {
             var formatCoordinates = Coordinates != null ? $"Reference Coordinates RA: {Coordinates.RAString} Dec: {Coordinates.DecString} Epoch: {Coordinates.Epoch}" : "";
             return $"FocalLength: {FocalLength}" + Environment.NewLine +
                 $"PixelSize: {PixelSize}" + Environment.NewLine +
-                $"ImageWidth: {ImageWidth}" + Environment.NewLine +
-                $"ImageHeight: {ImageHeight}" + Environment.NewLine +
                 $"SearchRadius: {SearchRadius}" + Environment.NewLine +
                 $"Regions: {Regions}" + Environment.NewLine +
                 $"DownSampleFactor: {DownSampleFactor}" + Environment.NewLine +